refactor: rename compute-attester and file architecture
