+++ attestations-collections/playground/polygon/hydra-s1-simple.ts
@@ -155,23 +155,6 @@ export const hydraS1SimpleBadges: BadgesCollection = {
   ],
 };
 
-export const pythia1SimpleBadges: BadgesCollection = {
-  collectionIdFirsts: {
-    [Network.Polygon]: 30000001,
-  },
-  badges: [
-    {
-      internalCollectionId: 0,
-      name: "[playground] Synaps Liveness ZK Badge",
-      description:
-        "[playground] ZK Badge owned by users that proved their liveness with Synaps",
-      image: "synaps_liveness.svg",
-      attributes: {},
-      requirements: ["Prove your Liveness with Synaps"],
-    },
-  ],
-};
-
 export const frontBadges: BadgesCollection = {
   collectionIdFirsts: {
     [Network.Polygon]: 0,
