+++ attestations-collections/staging/rinkeby/hydra-s1-simple.ts
@@ -111,23 +111,6 @@ export const hydraS1SimpleBadges: BadgesCollection = {
   ],
 };
 
-export const pythia1SimpleBadges: BadgesCollection = {
-  collectionIdFirsts: {
-    [Network.Rinkeby]: 30000001,
-  },
-  badges: [
-    {
-      internalCollectionId: 0,
-      name: "Synaps Liveness ZK Badge",
-      description:
-        "ZK Badge owned by users that proved their liveness with Synaps",
-      image: "synaps_liveness.svg",
-      attributes: {},
-      requirements: ["Prove your Liveness with Synaps"],
-    },
-  ],
-};
-
 export const frontBadges: BadgesCollection = {
   collectionIdFirsts: {
     [Network.Rinkeby]: 0,
